Make alerts more visible fix #79 + change the way notification permission is requested + Location search advices when no results fix #133 + Change location list top bar fix #70 + store data with decimals fix #138 + Add app to IzzyOnDroid + add icon to Fastlane + Update README

+++ app/src/main/java/org/breezyweather/common/basic/models/weather/Current.kt
@@ -21,7 +21,7 @@ class Current(
     val relativeHumidity: Float? = null,
     val pressure: Float? = null,
     val visibility: Float? = null,
-    val dewPoint: Int? = null,
+    val dewPoint: Float? = null,
     val cloudCover: Int? = null,
     val ceiling: Float? = null,
     val dailyForecast: String? = null,
